pyproject.toml
[project]
name = "fs-checking"
version = "0.1.0"
description = "Financial statement checking with multiple LLM strategies"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "pymupdf>=1.24",
    "python-dotenv>=1.0",
    "click>=8.1",
    "jsonpatch>=1.33",
    "pillow>=12.1.0",
    "playwright>=1.40",
    "markdown>=3.10.2",
]

[project.scripts]
fs-check = "scripts.cli:main"
fs-inject = "scripts.inject_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fs_checking", "scripts"]

[tool.hatch.build.targets.sdist]
include = ["src", "scripts"]

[dependency-groups]
dev = [
    "matplotlib>=3.10.8",
]
